requirements.txt
numpy==2.0.0
PyYAML==6.0.1
timm==1.0.7
torch==2.3.1+cu118
torchvision==0.18.1+cu118
tqdm==4.66.4
datasets